Add missing Lint checks
This changeset adds 43 lint checks that were available in lint but not
in IntelliJ; perhaps because some of them were using Java, and perhaps
others because they were added to Lint after the initial IntelliJ lint
integration.

There's also a unit test which runs and makes sure that all the issues
available to lint are also available in IntelliJ, except for a few
manually excluded issues (where IntelliJ already looks for the same
issue in its regular inspections.) The test spits out the
registrations to be added to plugin.xml, AndroidBundle.properties and
AndroidLintInspectionToolProvider.java.

(I also replaced some of the key names such that they are completely
determined by the corresponding lint issue id, and in alphabetical
order by issue id.)

Change-Id: Ide2f32008a352705896ee15c912c08734c148625