Add Android palette variations
This changeset adds a number of "palette variations"; these are variations
for the same XML view tag (such as <LinearLayout>) which initializes the
tag in different ways. In the case of LinearLayout we're offering both
a horizontal linear layout and a vertical linear layout. For text views there
are large, normal and small styles, and ditto for ProgressBars. And for
text fields there is a whole new category with various input types.

Change-Id: I3f32a399dec58b705b85200fa53e2d851b0b25b4