Switch from version names to API numbers
Generate the Android code names in the tooltips dynamically,
and store the API numbers in the layout. This allows us to
directly compare the required API level with the current
API level, so we no longer need to attempt to load all the
classes to see if they are available.

Change-Id: Ibc9e99ba2d6792702593d09de56cc78dc2c4b33c