Ensure that id's are added on <fragment>, not on <Space>, etc.
By default we want to assign ids to non layouts such as
Buttons and TextFields. There are some other exceptions to
the rule; we want ids on <fragments>, but not on <include>
and <Space> tags, etc.

Also, fix sequence id such that it's button, button2, button3,
etc instead of button, button1, button2, ...

Change-Id: I8d7e919174536e7fa2c87cc6a15b46b52e65e642