New project wizard.
Template-based new project wizard for IDEA. The infrastructure from
Eclipse was ported over and adapted to the new way of doing things.
This replaces IDEA's new project wizard entirely (it needs some
resource changes but not code changes to hook it up). It doesn't use
the IDEA new module wizard infrastructure, so it won't play nicely
with other new module wizards if you're using it as a plugin
in Ultimate Edition; we should decide how to deal with that.

Enable it by setting the android.newProjectWizard Java system
property. It also requires a new set of templates to reflect the
need for a build.gradle file and the new directory hierarchy.

As of right now it runs through a template and outputs a working
Gradle-based project. It doesn't actually import that project so
it doesn't do its stated function of actually creating a new IDEA
project, though what it outputs compiles and builds a working APK
from the command line.

The wizard doesn't yet put up the step to deal with the
compatibility library, and doesn't yet allow the user to create a
launcher icon in the flow. Input validation is somewhat implemented
but in a primitive state. The features to support the creation of
new templates are left out for now. Other than that most things
should hopefully work.

Change-Id: Id6019f4c5d5b17b3ee184958c984b40722108a14