Auto zoom the designer, and improve margin handling
First, this CL adds auto zooming to the android designer, such that
when you open a new layout, or when you switch orientations or device
type, the designer zoom is automatically zoomed (up to but not more than
100%).

Second, it improves the handling of margins such that the margins around
the design surface is up to 30,20, but less when the view is zoomed (and
the margins dynamically change as you're resizing the whole IDE view.)
It also removes some hardcoded margins that caused vertical scrollbars
even when not necessary.

Finally, it improves the scaled rendering code such that it doesn't
create an intermediate scaled image when zoomed it (scale factor > 1),
since in this case we don't need the careful thumbnail scaling code
(which resizes in multiple passes), and besides, it can use a lot of
unnecessary memory to keep a giant scaled rendering of a view where only
a small portion is actually painted in the viewport.

Change-Id: Ic375e279356491be85735a5fa476981a3be7917c