Add Margins class
This CL adds a new Margins class. This is used instead of a Rectangle
for margins, since Margins are logically different from Rectangles
(e.g. when converting coordinates, it is important *not* to treat x,y
as a location), and it also makes the code a bit cleaner (e.g.
you don't have to know that width=right or height=bottom.

Change-Id: I8e3a1fd5bebc3e7106ea4220f1c107dbfcd3c2f6