Handle dragging outside root bounds; retarget to root
If the root layout does not have its width or height set
to match_parent, the bounds of the layout may not extend
over the whole screen surface. If you drag over this area,
nothing happens; you have to target the bounds in the top
left corner.

This changeset changes this such that if the mouse is not
over any views, it will target the root layout instead.

Change-Id: Ie3ab32661d893a686d1b0127656b242b83c9e9b1