Launcher icon creator.
Adds a page to the New Project wizard to let the user create a custom
Launcher icon, based on the Asset Studio library. For now the wizard
page and image creation code are specific to launcher icons, but soon
this will be refactored to support multiple wizards creating other
icon types as well.

Change-Id: Ia9d16c8b3c1a0f11ca2a80badb7c07c0f9251026