New RenderService, and layout preview improvements
This changeset adds in a new RenderService similar to what we had in
ADT (a pull parser for XML based on the PSI infrastructure, such that
we can for example dynamically replace match_parent with fill_parent,
add padding etc), configurations and configuration listeners, and
various UI code to make the configuration chooser more similar to what
we had in ADT (menus including flags and language and region names,
etc).

One advantage of using the PSI structure as a basis for the layout
parser is that it makes it much more resilient to errors: even if the
user is typing in (incomplete) XML, only a fragment of the XML is
marked in error, and the layout render will generally succeed.

This CL also begins the work of unifying more of the XML based render
preview with the rendering used in the layout editor. In addition to
using the shared code to do better thumbnail zooming (by doing
repeated half size scaling operations) and drop shadows, it
synchronizes with the editor, such that you can click on the preview
to select the corresponding XML element in the editor, and conversely
when you move the caret around the corresponding view is highlighted
in the preview.

This work is still in progress; and sadly, it's not small and
isolated, which would be difficult to achieve with a change to
something as core as rendering and configurations.

While the configuration code originates from ADT, it's improved a bit
here; it has been fully decoupled from the configurations UI which
finally made it possible to write proper unit tests which is added
here.

Change-Id: Iea02b0b347170357ce49b7f6250a4213b9b0cc40