New render error handling
This changeset replaces the error list in the layout render preview
with an error panel near the bottom which is overlaid on top of the
rendering (and is partially translucent). The error output is
formatted HTML (similar to the code analysis details pane). There are
hyperlinks to perform actions (also similar to e.g. how suppress works
in the code analysis panel).

The existing fixes were ported (e.g. configure SDK, or change
compilation target from 1.7 to 1.6) and new ones were added, such as
actions to insert all missing layout_width or layout_height
attributes, checking missing class names for typos and offering to
replace them with the correct spelling, building the project, creating
a custom view class from an unknown tag, show javadoc, as well
as diagnostic help such as identifying missing style errors and
suggesting that the wrong theme might be chosen.

Change-Id: I4f903af43473bf7f4b2e7d5b090d1ff385cb6f45