logcat: Move list of filters to EditLogFilter dialog
The logcat view shows a list of filters that can be selected
in a listview to the left of the log panel. This CL moves it
out of the list view and into a combo box right next to the
console filtering options. The list of filters shown in the
combo box can now be edited within the Edit Filters dialog.

A subsequent CL will turn the device selection combo into a
list view occupying that space.

Change-Id: Ib59335c71d58d56c7b50f1d4bd7aa58a346f79ce