Persistent Project Resources
This CL adds a persistent project resource repository which records
the various resources (strings, layouts, drawables, etc) in the
project. This is used to for example handle layout rendering.

Until now, the project resources were loaded for each layout render.
Now, they are instead loaded once, and then kept up to date using PSI
listeners. The project resources instances (one for each module) can
be looked up on the Android facet.

In a followup CL, this will be rewritten to fully rely on PSI, both
for the initial population, as well as to perform incremental updates
(right now it simply uses the PSI listeners to record which files have
been added, deleted or changed, and then all the project resource
methods will lazily refresh themselves when queried.)

It will also need work to handle the new multiple resource directory
layout in the new build system, and the library projects shouldn't be
inlined into each project resource instance as they currently are.

This CL also adds ResourceResolver caching, tied to configurations.
This allows repeated renders to be faster.

Change-Id: Ic26f476d2abc79b1fa11d5e3ea69ca84a245dd20