Migrate layout editor to new render service
This changeset migrates the layout editor over to using the same
rendering service as the layout preview window, which sits on top of
the PSI model (and the persistent project resources). This means it
uses the same configuration bar and configuration state (so switching
to a device, orientation and theme in one also affects the other). It
also switches the error UI over to the same error overlay used in
render preview, with hyperlink suggestions.

This makes preview rendering work for ListViews and GridViews, as well
as <fragments>; this changeset also adds error panel hyperlinking of
<fragment> tags that do not have preview metadata such that you can
pick a designtime layout to be used.

Finally it removes the old render util code that is no longer used,
and various bug fixes here and there.

Change-Id: I51947dae70caafe049784f81509b44d28013f0a9