Handle file switching to best configuration file match
This changeset handles automatic file switching to the best
configuration.

If for example you are editing layout/foo.xml, and you also have a
layout-land/foo.xml, and you click on the flip-to-landscape toolbar
action, rather than rendering layout/foo.xml in landscape mode, it
will switch to editing the layout-land/foo.xml file instead.

In ADT this was handled by listening on configuration changes in the
UI, and if a better match was found, it would revert the UI and reset
the configuration data (such that we don't use orientation=land with
the prior file). Here however we instead have a new
ConfigurationAction class which handles this. Each configuration
simply implements a method to set its state on a given configuration,
and the configuration change action will first clone the
configuration, then let the action update its state (while listening
for affected configuration flags), then see whether a better file
match exists and if so switch to it (throwing away the incompatible
configuration), or if no better match exists it will call the update
method again, this time passing the real configuration.

Change-Id: I18f1f6d5cce35a67ac34edc13514d9c98cbe574c