Ensure that render caches are cleared when project changes
If you open a layout (such that it is initialized with a cached set of
framework and project resources), and then you create a new layout,
and finally you attempt to <include> the new layout from the first
layout, you get a render error.

The reason for this is that the cached configured project resources do
not include the @layout resource for the new layout. When I
encountered this and went to investigate how we made it work in ADT, I
was shocked to discover that we didn't handle this in ADT either!  The
above scenario doesn't work; it was however worked around for the
Extract Include refactoring with some manual refresh trigger calls.

The fix is simple: In the ProjectResource class, where we already
listen for project changes and update the project resource data
structures lazily, add a "generation" concept. Layouts which create
their own caches (e.g. project resources applied with a particular
layout) can also store which generation they based their cache on.

On later updates, they can check whether the project resources is at a
more recent generation, and if so, recompute the cached project
resources and resource resolvers.

Change-Id: I7e60a8cf5c9e03005cb3aebece5bfabae97a6ca9