Change language/region dialog to show full names and flags
This changeset updates the language and region editors in the new
resource file dialog such that rather than showing a textfield (where
the user needs to enter a language or region code), it shows a
filtered list with the known ISO language and region codes, along with
flags when available.

Plus unit tests for the LocaleManager which were missing.

Change-Id: I4a2ddfc3a52bf8e22733c022e28b9af3cb1d0854