New Theme Chooser
The theme drop down is no longer a drop down action; instead it's a toolbar button
which pops open a dialog where you can select themes by categories, and search.

The themes are now computed from the ResourceRepositories, so it's a lot faster.

Removed now defunct code, and cleaned up the Locale menu a bit.

Change-Id: Id4d8ec79a931b8639fd7d5fcb6aa0de6dd2f142c