Do not invoke android plugin code from core IDEA
The adt-branding module is part of the core IDE and cannot
refer to classes present in optional plugins such as the android plugin.
This CL adds an extension point in the core plugin that the android
plugin can perform to do Android IDE specific initializations.

Since this extension will only be available in the Android IDE, this
dependency has to be marked as optional and loaded only when it is
available.

Change-Id: I36c30235ce713ec8285f09c5a04d57a12e48d1af