Multi-configuration previews
This changeset adds basic multi-configuration preview support. Note
that it's only hooked up to the XML preview window, not the layout
editor, at this point.

It has a new tile-based layout which is more suitable for XML preview
than the layout editor, where the rendering does not need to be as
prominent since it is not directly edited -- so there is more room for
the previews. (The current implementation also does not support
rendering included layouts or manual preview lists.)

In addition to the multiconfiguration editing fixes, there are also
various fixes to related areas, such as configuration handling and
render read locks.

Change-Id: I5dd00ec385e8d1362d19577678e513e736b14fea