Fix bugs in rendering error panel
This changeset fixes a couple of bugs in the HTML render error panel.

First, when one or more views throw rendering exceptions that a view
is missing layout_width or layout_height, there is no record from
layoutlib of which view is missing it, so the error panel searches
through all XML tags and adds hyperlinks for all views found to miss
width or height. Unfortunately, in layouts which are relying on style
attributes to set sizes, this causes the error display to light up
with dozens of links, and clicking on them add redundant sizes. This
fixes takes advantage of the fact that we still have the resource
resolver from the rendering, so it uses that to look up each style
referenced by each element missing a size, and checks whether the size
is found there.

Second, it fixes a couple of formatting bugs where you could end up
with double commas (first, , second), and it uses a non breaking space
for list items to ensure that with wide custom view names we don't end
up with the list item bug on a line separate from the custom view
name.

Change-Id: I7fb5eae7cf22e63eb577da068bd6b0b7cb8d68e6