Add action to obtain a screenshot from the device
This CL adds a "Screen Capture" action to the Device Panel.
When invoked, it obtains a screenshot of the currently selected
device, and shows it in a screenshot viewer. The screenshot viewer
dialog provides the following actions:
  - Capture the screenshot again
  - Rotate the captured image
  - Frame the image with device artwork. Optionally include
    drop shadows or a screen glare effect.

The device artwork itself is a unmodified copy of whatever is
available at http://developer.android.com/distribute/promote/device-art.html

The JSON syntax comes from the same URL, but we can tweak this if
necessary.

Change-Id: Ib0760f798649fc9fe25c15a45cd802f892f80a9a