Add new lint checks and properly handle analysis scopes
This changeset adds in recently added lint checks as local inspections,
and removes some redundant declarations.

It also makes the lint bridge properly handle the Intellij AnalysisScope,
such that if you run inspections on a file or on a module, lint only checks
those files/modules rather than the whole project.

Finally it fixes an NPE in the Lombok PSI bridge.

Change-Id: Iae186f736be0c4e743ac13058af9b65e8beecb80