Fix configuration matching
This CL fixes a couple of issues related to the configuration matcher
which tries to adapt the current persisted configuration to the currently
selected configuration.

First, it adds in a version qualifier, which fixes an issue where layouts
in a -vNN folder could not be adapter (it would fall back to the full
configuration selection).

Second, it sorts the current project locale first such that it is picked
among all the matching locales when testing each.

And third, add a fallback device to the front of the list such that we don't
pick the N1 when the most recent devices.xml devices aren't installed in
the SDK folder yet.

Change-Id: I9f0b88e0bee91548ba7c4f8fe4f07c64562bd7a8