Attempt to automatically pick an Android SDK if needed
When loading in a module, if no Android SDK is associated with
it (which can happen if for example you have updated your Android
SDK and the version number has changed), it will automatically
pick one if another one is found, and store that selection.

Also make error assertions in the renderer provide a bit more
context, since I suspect many of the user reported stack traces
which just said "Render Exception" were really because no SDK target
was configured upon editor open.

Change-Id: I55ca14ca1258d60403e4203839443eef0dd9275e