Handle module dependencies in Gradle projects
This adds a temporary workaround for the utility mehod which returns all
dependent AndroidFacets from a Gradle project. Longer term the ModuleRootManager
should be returning these, but there are apparently deliberate reasons why this
is not the case yet. Among other issues, this should make it such that the
editor can resolve resource references in dependent modules.

Change-Id: I4707b3b75a4dad24ab04156e3924f932c344d8cb