Improve resource directory lookup in Gradle projects
This changeset hooks up the resource directory lookup to the gradle
model's resource directories such that for example the editor can resolve
resource references in other folders than the base res/ folder.

This also fixes the code which determines whether an XML file is really
a resource file (for example for UI preview), but note that actually
properly rendering resources from flavor directories requires a lot more
work which is handled in a CL dependent on this one.

(There are some other miscellaneous minor fixes in here as well.)

Change-Id: Id306951f2cbe49574cc719be5a8656a282d19931