Gradle IDE resource support
This changeset adds support for Gradle project resources. In particular, it
replaces the older project ResourceRepository with a new implementation based
on the res2 ResourceRepository, which is built on the new build library for
handling resource merging. It migrates all the client code which was relying
on APIs only present in the old ResourceRepository (ResourceFile and ResourceItems)
over to the new way of doing things, and starts the migration work of hardcoded
resource directory references to using a full resource list. It also adds a
delegating project resource for use in multi-module projects.

It also starts the work of adding listeners on the gradle project model and
deferring layout preview and layout editing until the full model has been synced,
as well as refreshing when the variant changes.

Change-Id: Icdd5efed0e850ec79b485dcc8bb62f6b525f5263