Fix configuration matching
This CL fixes bugs in configuration matcher (which for example meant
you could open up a layout in landscape orientation and it wouldn't
switch to a better match from the -land folder.

The root culprit is that after switching from ResourceFile to
VirtualFile we now pass in file extensions into the matching code,
which isn't right; we shouldn't take file#getName but use
ResourceHelper#getResourceName instead.

Change-Id: I1aa8c6ee8c11fdb442f2caab20e600adf2aa4204