56339: Improve render error message handling
This CL fixes a few issues around render error messages.

First, when there are many repeated error messages of the same type
(for example, in one sample project I broke the parent theme
reference, and as a result got 40 messages stating that they couldn't
find various items in the theme) these are shown just a single time,
along with a "<count> similar errors not shown" message. Similarly,
for render fidelity errors, only the first 3 fidelity warnings (which
are already collapsed per type) are shown.

Second, it fixes a couple of bugs; suppressing fidelity warnings for a
given type was broken (by an invalid assertion), and the theme
selection dialog did not work correctly for project themes.

(Issue 56339 asks for the render error overlay panel to be resizable;
this CL does not fix that, but for now reduces the need for it to be.)

Change-Id: If9eaf278b4499791c2d9b7389cb76465064adc01