Look up original source files from copies in build/
This CL makes the error handler look up the original source
files from error messages pointing to files in the build/
directory. This works by using comments left in the copied
files to point back to the originals. For the merged values.xml
file it's slightly tricky since there's no information about
the original line numbers, so it uses a heuristic to search
for the relevant line if the error message contains clues about
where the error is.

Change-Id: I7b6302bd8e74b9fce287e33b5991103bbbe802e1