Fixes to context sensitive launch configurations.
This CL enables users to right click on an Activity class and
create a launch configuration that will launch that activity.

In addition, it overrides the findExistingByElement() method so
users don't see the option to create redundant launch configurations
when a suitable one already exists. This is done for both activity
and test launches.

Change-Id: I4e1ec7c92234dec9b2c3eb446ff3eadf591898eb