Handle module root changes on import and delayed initialization
This changeset makes the transitive project resources (ModuleSetRepository)
handle the case where (1) it's initialized at startup, before the gradle project
is ready, so it doesn't find out about library dependencies, and (2) a subsequent
gradle project sync, where the module roots changes. In both cases it now will
always use a delegating repository (even when the number of children is 1) such
that it can always switch to more children, and it recomputes its set of children
whenever the module roots change.

Change-Id: I7912d619d54177f68335ffe561fe7501c0f507a7