Display flavor and build type resource items in popup documentation
This changeset enhances the javadoc popup for resources (strings,
dimensions, drawables, etc) such that it not only displays the values
across resource qualifiers, but also across build types, flavors and
libraries.

It will display the highest precedence match first. Any subsequent
matches (per resource folder) are shown in strikeout style to make it
clear that the item is masked by a later resource folder. (The items
are sorted in reverse folder order, meaning it will display build
types first, then flavors, then the main source set, then any matches
in source sets that are *not* selected, and finally matches in the
libraries.) The selected flavor source set names are shown in bold.

The javadoc for resources will now also work in value files (meaning
you can invoke it <string name="here">.../</string> in a values file
to see the other definitions.

To support the above feature the resource repository support was
improved a bit; there is now a ResourceFolderRegistry responsible for
managing resource repositories for arbitrary resource folders (used to
for example query resources in folders that aren't part of the current
variant.)  Similarly, repositories now have names (so we can display
for example that a given item comes from library X), and most
importantly, some cleanup in the way resource folders are managed to
avoid some assertion errors reported against Studio 0.1.9.

Change-Id: Ieb00c8a7fecc08ecce261e4f7ad57b6d11c6f900