Improved reporting of errors when importing a Gradle project.
When an error happens while importing a Gradle project, the real cause of the error is wrapped in
many layers of exceptions, resulting in an error message that is completely useless. Users have
to go to the log file to see what's going on.

This CL does 2 things:
- Returns the real cause of the error.
- When possible, replaces the message of the error with an user-friendly one.

In this CL we are only improving the message of the error that occurs when using an old version
of Gradle. We will improving error messages as we find more not-so-user-friendly ones.

Fixes issue https://code.google.com/p/android/issues/detail?id=57121

Change-Id: I3e3fe834b2c262635f4f8b259d250c86b4b33384