Fixed configuration errors when creating a new Gradle-based Android project.
The problem is that AndroidGradleProjectComponent#projectOpened is called
right after the project is created, when the facets have not been created,
instead of when the project is *actually* opened. The side effect of this
behavior is project settings not being set and event listeners (e.g.
ModuleListener, BulkFileListener) are not registered and functionality
like renaming a module's folder will not work until the IDE is restarted.

The solution is to manually invoke AndroidGradleProjectComponent to setup
Gradle-based Android projects.

Change-Id: I241bbae476eaa181ff1ce06bc34a5d8c16d5862d