In Add Translation dialog, require at least one string
Without this, you could Add a Translation for a language, just
hit OK, and end up with a scenario where the UI thinks you've
added a new language, but the resource repository does not, and
the locale menu is confusing. With this fix, you have to enter
at least one string (which is what the resource repository requires
in order to consider the language to be present), or cancel out.

Change-Id: Ib2e340185f8de44d806a681d225190a4331e9e05