Removed gradle-0.5.0.jar now that we don't use RMI for project import.
Because of RMI, we had to have this jar so RMI marshalling and
unmarshalling could work. Since this change affects TemplateMetadata (it
was using a class in the removed jar to get the current supported Gradle
version,) I updated the supported version to "1.7" so new projects will
have this value in the Gradle wrapper.

Change-Id: I0e448461f166dc4931cba722258e968b9bf5af51