Fix missing resource layout rendering errors
With the new PSI based rendering repository, the resource repository
is updated incrementally. However, there are cases where the
incremental analysis fails, for example if the PSI event notification
does not contain enough details (this is documented in the unit
tests).  In those cases, the incremental updater falls back to
updating the whole file, and this is delayed until the next dispatch
thread update in order to buffer up lots of small calls.

It turns out the layout editor also listens to the PSI hierarchy for
notification of when to update itself. However, when you for example
add a new resource, this update would run synchronously and perform
the layout render *before* the asynchronous resource scan had been
performed.

The fix is easy: for the PSI layout update listener, defer rendering
one iteration. That should always pick up any pending resource
updates.

Change-Id: Id6807ba2525e0d0bde308c33f1a3c388b8897729