Added link to Gradle docs to set up HTTP proxy settings.
With the new 'in process' Gradle project import, we cannot longer
pass IDEA's proxy settings to the intermediate process (this process
does not exist.) Instead, users need to configure Gradle's own
HTTP proxy settings. In this CL we provide a link to the Gradle
docs regarding these settings and modified the text of the
user-friendly error to make it clear what to do.

I also removed the configuration of the RMI project import, since
we no longer need it.

Change-Id: Ie5429519231936c9625784c66267e4f37e3e246c