Show exception in render errors
The method which appends exceptions to the render error looks at the
stack frames and only dumps out "relevant" stack frames, down to the
layoutlib entry point. However, the method is called from two places,
and in the place where we can't instantiate the class the exception
header is already printed; in that case we should display basic frames
such as problems in the android view code.

Change-Id: Ic456b86bb423f5ad0192c13f067f5400ba19e3fb