Importing project now populate "Gradle" view with available tasks.
Fixes issue https://code.google.com/p/android/issues/detail?id=59204

As part of this CL, I merged ProjectResolver into
AndroidGradleProjectResolver. There was no need to have 2 classes
for importing a project.

In Gerrit there will be a lot of "new code" blocks. All of them,
except for method "addModuleTasks" were moved from
ProjectResolver. The method "addModuleTasks" is new, and it is the
one that populates the tasks of the imported project.

Change-Id: I159c0bb2f9c00db9849a1e610adf5e1dada4c23a