logcat: Handle spaces in tags properly
The logcat formatter is used to print out a logcat message for
display in the logcat view, but is also parsed back again so as to
assign the message an appropriate color.

This mechanism assumed that a logcat tag cannot have spaces, and that
a colon character delimited the tag from the message.  However, that
isn't really correct since the tag and the message could have any character.

This CL fixes this issue by using an infrequent Unicode character as
the delimiter. The character is currently \ufe55 (small colon) which still
looks like a colon but is just unlikely to show up in a tag or a message.

Change-Id: I436fa62208570668652553d67f360f6cbb639afa