Fix issue: "Clear logcat before launch" doesn't work.
This fixes https://code.google.com/p/android/issues/detail?id=57041

AndroidRunningState assumed that the logcat view could be directly
accessed from the toolwindow's content. This assumption was broken
when we moved logcat to be a panel inside the DDMS toolwindow.

To fix this, we simply store a reference to the logcat view in the
DDMS toolwindow as well.

Change-Id: Ie14581099e4a81061e9e4c1621562807d0a34b8d