Fix gradle output parsing
This CL fixes two bugs in the code to process gradle output.

First, it turns out that gradle can sometimes use file: and file:// URLs
in the values.xml files recording the original source locations. Since
our expression matchers did not expect that, and they only pick up file
names that they determine to exist, this meant the error location was
ignored.

Second, there was a bug in the search code (which looks for the original
position search marker), which meant it might not find the right source
marker and could point to the wrong file.

Change-Id: Ia7dcbedbf1c4c30108b626bb80f6d4735e3a9a5f