Improve layout bitmap caching; reduce memory usage
There appears to be a leak in layoutlib related to the Bitmap native
delegate which has been there for a long time:
https://code.google.com/p/android/issues/detail?id=59498

However, the problem seems to have worsened in Studio relative to what
we observed in ADT. It looks like the reason for this is that in ADT,
the layout lib bitmap cache was cleared more often than in Studio. In
ADT, the bitmap cache was only cleared on .png file changes and on
render target changes, whereas in Studio it was cleared on every
render.

This changeset makes Studio update the cache for the same events as in
ADT, which should reduce the number of cache misses in the bitmap
cache, which in turn should reduce the number of bitmap delegate
instances, which in turn should reduce the impact of the leak.

Change-Id: I42702d95c708a7b1eacf33b7f6d517be901890d7