"Sync Project with Gradle files" action is now always visible.
Before this CL, this button used to be visible only for Gradle-
based Android projects. This turned out to be not very useful,
especially when a project import failed before the
"Android-Gradle" facet was applied to the project. In this case
users could not sync the project after fixing whatever was
the cause of the project import failure.

The Sync button is always visible in the main toolbar in
Android Studio only.

If the project is not a Gradle-based project, Studio shows the
appropriate message in a notification balloon.

In this CL I also refactored how notifications are built and
displayed, to the code more reusable.

Fixes issue
https://code.google.com/p/android/issues/detail?id=56373

Change-Id: I1cc8916b20b23d1ddf7be2e93a54b853c218c5ee