Fixed distribution type when importing newly created project.
In this CL, we change the distribution type from WRAPPED to
DEFAULT_WRAPPED. WRAPPED does some funky stuff to determine
the location of the wrapper, i.e. creates a Gradle script
on the fly and calls the "wrapper" task on it (an extra
connection to the Gradle Tooling API.) This is not necessary
for projects created with the "New Project" wizard, because
it creates the wrapper in the default location.

Change-Id: I0fbe61cf3747e0509decd0fc1d81b8800c14af2d