Speed up editing in layout XML files with render preview
This change makes the PSI listener, which listens for edits and
triggers the layout preview to kick in after a brief delay, ignore a
lot of edits, such as any whitespace-only edit, inserting new
attributes (before they have values), edits in a comment, and so on.

This makes it less likely for the editor to stutter and think when you
type in new attributes and invoke code completion and you're waiting
to see the possible alternatives.

This CL also cleans up some code that is no longer necessary; the code
was there to detect when other resources (strings, styles etc) had
changed. That's not needed now that we have a dedicated resource
repository which is already monitoring PSI changes and updating the
repository along with change flags for the configuration.

Also fix some warnings.

Change-Id: I635becae94ca475bb37187ce2d14865158011dbc