Expand elided frames in stack traces
Java errors that have a non null cause don't print out the full frames
of the cause. The frames that match the outer exception are not
displayed and are replaced by a "...N more" string in the stack trace.

This CL replaces that line ("...N more") with the actual frames from
the outer stack trace, and adds a folder that folds those line back
into a single line. This allows users to expand and see the full trace
while retaining the simplicity of just seeing "...N more" by default.

This works well except in the case where the expanded frame is also
part of the default list of frames folded by the IJ console. (This list
can be viewed in Settings | Folding). In such a case, rather than the
full expanded list being folded, only a part of it might be folded.

Change-Id: I7d001e3d22b9afdd4471f605c0d29622accfb716