logcat: Remove redundant "Clear log" button.
According to https://code.google.com/p/android/issues/detail?id=60893,
there are two ways to clear the logcat console - one uses the "Clear Log"
action that explicitly clears logcat on the device, and the other uses
the "Clear All" action which is part of every IntelliJ console. The
latter clears the console, but the impact is transient - switching back
and forth from another toolwindow will show the old log content again.

This CL:
 - Removes the redundant "Clear Log" button
 - Overrides the "Clear All" action to behave like the "Clear log" button
   used to behave i.e., it clears the logcat buffer on the device.

This is performed by implementing the ConsoleActionsPostProcessor
extension point which allows us to customize the list of actions shown
for a console view.

Change-Id: I670c0cac0cd81776163c67919871507aca61bb08