60879: Handle padding and margins in RelativeLayout
The RelativeLayout support in the layout editor did not properly take
margins and padding into account, if set on the RelativeLayout itself,
which meant that after dragging views would jump slightly to their
real padded positions. This changeset fixes that; in addition properly
accounting for the margin and padding rectangles, it also draws the
padding rectangle if non-zero. (In a followup CL I will make sure the
various other layouts also handle padding properly.)

There was also a bug in hor view coordinate computation which meant
that the margins would not be computed correctly when the view was
zoomed; this is fixed.

Finally, I renamed the Margins class to Insets, now that it's used not
just for margins but for padding as well. Also fixed various warnings.

Change-Id: Ia532c3d938dbe176ebf8321fb6f32a00379cef9a