Change LinearLayout drag handling to only use gravity for center/right
The LinearLayout drag handler in the layout editor, for a horizontal
LinearLayout, used to use the vertical position to set the gravity
attribute, to either top, center, bottom, or fill_parent.  (The last
is not a gravity but a height, meaning that it would set the height to
fill, and set the height to wrap_content for the other gravities).

Setting the gravity explicitly to top is not good; it turns off
baseline alignment in the linear layout, which is normally what you
want for a horizontal LinearLayout.

This changeset changes this, such that we no longer set the gravity if
it is left (in a vertical layout) or top (in a horizontal layout).

It also turns off the fill_width handling, which was a bit asymetric
with the rest. In the layout editor we now have layout actions which
makes it easy to toggle this, and in a subsequent CL I'm going to work
more on weight handling and fill handling.

Change-Id: I72d5799ebb19ff09433362665079eb7509076418