Custom renderer for ArrayMap objects.
This CL adds a custom renderer that is used by the debugger
when displaying ArrayMap objects.

ArrayMap implements the Map interface, but its entry set does
not support the toArray() method. As a result, the default Map
renderer just shows an UnsupportedOperationException while
examining the contents of an ArrayMap from within the debugger.

Change-Id: Ib044ff4054b072333141a951508e14eae38a17bc

Conflicts:
	android/src/com/android/tools/idea/startup/AndroidStudioSpecificInitializer.java