Android launch: Prompt user on INCONSISTENT_CERTIFICATES error
Installing an application on a device can fail if it is already
installed on the device, but with a different signature.

This CL detects this case and prompts the user whether the installation
should be tried again after uninstalling the existing application.

Fixes https://code.google.com/p/android/issues/detail?id=61108

Change-Id: I6218bd9b5c36274f1f9267a214c331292ec0360e