Now we prompt user to sync project is path of APK not found.
According to this issue
https://code.google.com/p/android/issues/detail?id=59018
and G+ comments, this error (APK's local path not found)
happens mostly after upgrading Android Studio.

The fix is simply to sync the project again, but it is
not obvious. This CL shows a balloon with a hyperlink
to sync the project.

Also this CL cleans up some code.

Change-Id: I3eb40c9e6d668c0436fe2feba238b04f611c618f